requirements.txt
Flask
Flask-Cors
selenium
webdriver-manager
beautifulsoup4
vaderSentiment
pymongo
python-dotenv
